Replace Quirk define with a quirks mode setting. The settings of the Quirks mode property are as follows:
detect - automatically detect if a quirks mode by scanning the document for the appropriate !DOCTYPE declarations.

standards - for documents known to be designed around the standard specifications
such as XHTML documents or something with a
"X-UA-Compatible	IE=Edge,chrome=1" HTTP header.

quirks - for documents known to work only with a quirks mode.